Renamed: wrong namespace name to correct Updated: phpunit to version ^7.0
