+++ composer.json
@@ -16,7 +16,7 @@
     "symfony/dependency-injection": "^4.0"
   },
   "require-dev": {
-    "phpunit/phpunit": "^6.5",
+    "phpunit/phpunit": "^7.0",
     "php-coveralls/php-coveralls": "^2.0"
   },
   "autoload": {
