+++ tests/TypeTest.php
@@ -16,8 +16,8 @@ use PHPUnit\Framework\TestCase;
 use Sauls\Component\Helper\Exception\ConverterNotFoundException;
 use Sauls\Component\Helper\Exception\InvalidTypeConverterException;
 use Sauls\Component\Helper\Stubs\DummyObject;
-use Sauls\Component\Helper\Stubs\Ensurers\CustomTypeAConverter;
-use Sauls\Component\Helper\Stubs\Ensurers\CustomTypeBConverter;
+use Sauls\Component\Helper\Stubs\Converter\CustomTypeAConverter;
+use Sauls\Component\Helper\Stubs\Converter\CustomTypeBConverter;
 use Sauls\Component\Helper\Stubs\JsonSerializableObject;
 use Sauls\Component\Helper\Stubs\StringObject;
 use Sauls\Component\Helper\Stubs\TraversableObject;
